Comment out Detect high network traffic
detect high network traffic commented out so the directional depth of field filter and other creative depth filters can be used in Star Citizen alpha for creative screenshots and videos. Star Citizen is not a competitive FPS game with smoke screen bombs so enabling the ability to use depth of field filters should not be a cheating risk for Star Citizen; however, this repo should not be used for any malicious usage or intended usage to cheat or bypass any security measures of any program. Such malicious use is strictly forbidden. This repo fork is solely intended for creative content creation use only.